separate file for content of link
